requirements.txt
psycopg2-binary
openai
requests
python-dotenv